+++ scripts/winesapos-convert.sh
@@ -91,7 +91,7 @@ Include = /etc/pacman.d/chaotic-mirrorlist" | sudo tee -a /etc/pacman.conf
     cd ..
     sudo rm -rf yay
     # GuestSneezeOSDev: Balena Etcher time
-    export ETCHER_VER="1.19.21"
+    export ETCHER_VER="2.1.3"
     curl --location "https://github.com/balena-io/etcher/releases/download/v${ETCHER_VER}/balenaEtcher-${ETCHER_VER}-x64.AppImage" --output /home/"${USER}"/Desktop/balenaEtcher.AppImage
     chmod +x /home/"${USER}"/Desktop/balenaEtcher.AppImage