+++ scripts/winesapos-install.sh
@@ -587,7 +587,7 @@ echo "Installing sound drivers complete."
 
 echo "Installing balenaEtcher..."
 # Etcher by balena.
-export ETCHER_VER="1.19.25"
+export ETCHER_VER="2.1.3"
 curl --location "https://github.com/balena-io/etcher/releases/download/v${ETCHER_VER}/balenaEtcher-${ETCHER_VER}-x64.AppImage" --output "${WINESAPOS_INSTALL_DIR}"/home/"${WINESAPOS_USER_NAME}"/Desktop/balenaEtcher.AppImage
 chmod +x "${WINESAPOS_INSTALL_DIR}"/home/"${WINESAPOS_USER_NAME}"/Desktop/balenaEtcher.AppImage
 echo "Installing balenaEtcher complete."