+++ scripts/winesapos-tests.sh
@@ -1078,7 +1078,6 @@ pacman_search_loop \
     cloud-guest-utils \
     curl-static-bin \
     crudini \
-    dtrx \
     firefox-esr \
     hfsprogs \
     mbpfan-git \