+++ CHANGELOG.md
@@ -6,7 +6,6 @@
 - Remove broken 'reiserfs-defrag' package.
 - Add support for the pifs (πfs) file system.
 - Change GE Proton version from GE-Proton9-25 to GE-Proton9-27.
-- Change Proton Sarek version from Proton-Sarek9-25-async to Proton-Sarek9-27-async.
 - Add support for the s3fs file system.
 - Add dual-boot support with Ubuntu.
 - Add dual-boot support with Fedora.
@@ -22,6 +21,7 @@
 - Change 'asusctl-git' package to use the 'asusctl' package instead.
 - Remove Xorg support.
 - Change file system commit time from 600 seconds to 300 seconds.
+- Remove Proton Sarek.
 
 ## 4.3.0 - The Wayland Update